requirements.txt
alembic==1.4.3
click==7.1.2
digi-xbee==1.3.0
dnspython==1.16.0
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
Flask-Threads==0.0.1a0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
monotonic==1.5
PyJWT==1.7.1
pyserial==3.4
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.2
python-socketio==4.6.0
six==1.15.0
SQLAlchemy==1.3.19
SQLAlchemy-serializer==1.3.4.4
srp==1.0.16
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
psutil==5.7.3